requirement.txt
numpy==1.15.4
Pillow==5.4.0
PyGLM==0.4.8b1
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
PyOpenGL-Demo==3.0.0
